Modify PaymentDetailsService for using MoneyAmount

+++ HappyTravel.Hiroshima.Common/Models/Accommodations/Rooms/CancellationPolicies/Policy.cs
@@ -20,6 +20,6 @@
         /// <summary>
         /// In nights number or percentage
         /// </summary>
-        public double PenaltyCharge { get; set; }
+        public decimal PenaltyCharge { get; set; }
     }
 }
